+++ src/PuliApplicationConfig.php
@@ -334,11 +334,11 @@ class PuliApplicationConfig extends DefaultApplicationConfig
                     return new PackageCommandHandler($puli->getPackageManager());
                 })
 
-                ->beginOptionCommand('add', 'a')
+                ->beginOptionCommand('install', 'i')
                     ->addArgument('path', Argument::REQUIRED, 'The path to the package')
                     ->addArgument('name', Argument::OPTIONAL, 'The name of the package. Taken from puli.json if not passed.')
                     ->addOption('installer', null, Option::REQUIRED_VALUE, 'The name of the installer', InstallInfo::DEFAULT_INSTALLER_NAME)
-                    ->setHandlerMethod('handleAdd')
+                    ->setHandlerMethod('handleInstall')
                 ->end()
 
                 ->beginOptionCommand('list')